application/backend/requirements.txt
# ============================================================
# AdviceRoom Backend — Dependencies
# ============================================================

# --- Web Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# --- MCP ---
fastmcp>=2.0.0

# --- HTTP Client (async) ---
httpx>=0.28.0

# --- S3 Storage ---
boto3>=1.35.0

# --- Redis ---
redis>=5.2.0

# --- YAML Config ---
pyyaml>=6.0.2

# --- Auth (JWT) ---
pyjwt[crypto]>=2.10.0
cryptography>=44.0.0

# --- Utilities ---
python-multipart>=0.0.18

# --- Dev / Test ---
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
